+++ src/dinojump/util/Account.java
@@ -1,4 +1,4 @@
-package util;
+package dinojump.util;
 
 public class Account
 {
